requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
cryptography>=42.0.0
aiosmtplib>=3.0.0
aiohttp>=3.9.0
tzdata>=2024.1
